Move common asset config to generator
It is collection of sidecar configurations, used by the individual
operator generators.
