+++ assets/overlays/aws-ebs/aws_ebs.go
@@ -7,8 +7,8 @@ import (
 	opv1 "github.com/openshift/api/operator/v1"
 	"github.com/openshift/csi-operator/assets"
 	"github.com/openshift/csi-operator/pkg/clients"
-	"github.com/openshift/csi-operator/pkg/config/common"
 	"github.com/openshift/csi-operator/pkg/generator"
+	common "github.com/openshift/csi-operator/pkg/generator/common"
 	"github.com/openshift/csi-operator/pkg/operator/config"
 	"github.com/openshift/csi-operator/pkg/operator/config/defaults"
 	"github.com/openshift/library-go/pkg/controller/factory"
