+++ pkg/generator/common/base_assets.go
