+++ pkg/generator/common/port_registry.go
